fix: cast the ram limit to int

+++ bigcodebench/evaluate.py
@@ -286,9 +286,9 @@ def main():
     parser.add_argument("--samples", required=True, type=str)
     parser.add_argument("--parallel", default=None, type=int)
     parser.add_argument("--min-time-limit", default=1, type=float)
-    parser.add_argument("--max-as-limit", default=128*1024, type=float)
-    parser.add_argument("--max-data-limit", default=4*1024, type=float)
-    parser.add_argument("--max-stack-limit", default=5, type=float)
+    parser.add_argument("--max-as-limit", default=128*1024, type=int)
+    parser.add_argument("--max-data-limit", default=4*1024, type=int)
+    parser.add_argument("--max-stack-limit", default=5, type=int)
     parser.add_argument(
         "--check-gt-only", action="store_true", help="Check the groundtruth"
     )
